Bug fix: use dispatcher to apply theme
